Refactored the EncryptionService
