+++ dao/src/main/java/com/hashmapinc/server/dao/encryption/EncryptionServiceImpl.java
@@ -14,24 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.server.dao.service;
+package com.hashmapinc.server.dao.encryption;
 
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
 
 import javax.crypto.Cipher;
 import javax.crypto.spec.SecretKeySpec;
 import java.util.Base64;
 
+import static com.hashmapinc.server.dao.model.ModelConstants.ALGORITHM_STR;
+import static com.hashmapinc.server.dao.model.ModelConstants.UTF_8;
+
 @Slf4j
-public class Encryption {
+@Service
+public class EncryptionServiceImpl implements EncryptionService{
 
-    public static final String UTF_8 = "UTF8";
-    private static final String ALGORITHM_STR = "AES/ECB/PKCS5Padding";
-
-
-
-    public static String encrypt(String plainText , String key) {
+    public String encrypt(String plainText , String key) {
         try {
             if(plainText == null)
                 return null;
@@ -39,15 +39,14 @@ public class Encryption {
             SecretKeySpec secretKey = new SecretKeySpec(key.getBytes(), "AES");
             cipher.init(Cipher.ENCRYPT_MODE, secretKey);
             byte[] cipherText = cipher.doFinal(plainText.getBytes());
-            String encryptedString = new String(Base64.getEncoder().encode(cipherText), UTF_8);
-            return encryptedString;
+            return new String(Base64.getEncoder().encode(cipherText), UTF_8);
         } catch (Exception e) {
             log.trace(e.getMessage());
         }
         return null;
     }
 
-    public static String decrypt(String encryptedText , String key) {
+    public String decrypt(String encryptedText , String key) {
         try {
             if(encryptedText == null)
                 return null;
@@ -55,8 +54,7 @@ public class Encryption {
             SecretKeySpec secretKey = new SecretKeySpec(key.getBytes(), "AES");
             cipher.init(Cipher.DECRYPT_MODE, secretKey);
             byte[] cipherText = Base64.getDecoder().decode(encryptedText);
-            String decryptedString = new String(cipher.doFinal(cipherText), UTF_8);
-            return decryptedString;
+            return new String(cipher.doFinal(cipherText), UTF_8);
         } catch (Exception e) {
             log.trace(e.getMessage());
         }
